requirements.txt
streamlit
tensorflow
numpy
opencv-python-headless
google-generativeai
groq
pillow
python-dotenv
requests
streamlit-webrtc
av